+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # CHANGELOG
 
-## 0.0.1
-First Version
+## 0.0.2 - 2021-08-22
+* HueBridge: Added ability to generate a username (token)
+* Added simple CLI: `phuey -h`
+  - Currently only supports one action: `create-username`
+
+## 0.0.1 - 2021-08-21
+* Initial Release